cleanup/refactors suggested by @ye-lou review
